add getCroppedCanvas method
This method returns a cropped area from the canvas in raw form.

The use of a primative for-loop rather than the preffered forEach or
map methods is for performance reasons: in the vast majority of cases a
value-increment loop is faster. Since all horizonal pixels can be set in
a single call, this loop needs only as many iterations as the height of
the cropped area but we can't be sure how often a client may choose to
crop the canvas.