guard against accessing canvas data while not synced
Without this, it was possible that a new pixel was placed before we
had canvas data. This would cause a crash due to a failed attempt
to get the previous color on the canvas. It's unclear if these
events should be dropped as they are here, or buffered until the data
is synced again.