improve handling of websocket
This is mostly comprised of three things:
- Buffering pixels until canvas data is loaded
- Breaking `setupListeners` apart
- Generally ensuring sync state is maintained better

`setupListeners` was a problematic function — it was designed to only
be called once. This would probably be fine if it was only called in
the constructor, but it was not. Secondly, the nature of binding it to
the events handler introduces an uncertainty with data state when an
external handler is invoked. Now, all events are processed internally
first before the event is even emitted.

Further, `setupListeners` was responsible for the gradual change of the
heatmap. This uses the metadata to determine the rate — something which
may change on any given server restart. This means that calling this
once is also insufficient — it must be called once every time the
metadata is set. It must also remove the previously set interval before
doing so as not to stack the “cooling” effect.