replace got with node-fetch
While I'm sure there's nothing wrong with got, fetch is a more standard
and consistent approach. I would also be lying if I didn't mention that
this is in part because discord.js uses node-fetch and, since fairy —
the sole dependant of this package — relies on discord.js, this change
allows consolidation of dependencies.