replace pngjs with sharp
Sharp is a bit more flexible, though it would be dishonest to not
mention this is to minimize dependencies in dependants on this library.

This comes with some API breaking changes, but only for those that used
clearly internal functions anyway.

The good news is that now snapshots can be saved in other formats than
just PNG. This can be done by changing the file extension provided to
the save methods.