fix heartbeat
Heartbeat has apparently been broken since its implementation. It's a
good thing so few people rely on this library right now because that
would have been a massive L. Even now, it's still a pretty embarrassing
failure to have let happen.

The actual problem was that it relied on the server sending pings, which
the server does not. This results in an expensive reconnect every 31
seconds.

This fix also comes with a refactor of websocket code in general. There
are a couple of API breaking changes:
- `closeWs` → `disconnect`
- `restartWs` removed

`disconnect` actually waits for the connection to disconnect — unlike
`closeWs`.