check property initialization
I'm guessing this is a leftover from where I got this compiler and
linter config from — gjs. Since GObject is a bit fucky in non-C
languages, having the compiller be happy about readonly properties is a
bit harder. For this reason, it's better to ignore the error there.

We are not boung by GObject's machinations here and could do with being
told when we've missed something.