add template handling
This comes with a native module in rust lifted from fairy. In fact,
basically all of this is from fairy. The hope is to provide a basis for
library users to build template handling on, so taking from the most
prominent (only) template handling application using this library seems
to make sense.

The native module is currently only distributed as a x64 linux binary,
but the hope is to distribute several compiled version according to
platform and os. In the event that the module fails to load, a slow JS
fallback should polyfill, though I haven't tested this code for full
compatibility like I probably should have.